New configuration engine
I reversed the direction of the concept: previously the config parser
would keep track of the key-value pairs as a dictionary. config.c would
then query the values based on the keys of the options. This is the
python config parsing model. It causes problems because I plan to
implement an @inherit directive, which relies on the available options,
as well as on the top-to-bottom parsing order.

Now the options are declared to the config parser in advance, in the
form of option callbacks: a key is bound to a callback with the value
as one of the callback's arguments. Aside from being reqiured in order
to implement @inherit, it should also perform better.