Split edit.c into edit.c and mode.c
This is to facillitate the implementation of more modes, and scripted
modes in the future.

The plan for now is to first implement a few modes (parameter selection,
block selection, token selection) in C, that only work properly when
editing C. Later these modes will be implemented through Lua scripts,
and then support for other languages can be added easily too.

Language independent selection modes will remain hardcoded in C.