Add binary tree implementation
This is based on the http://github.com/mirek/rb_tree red-black tree
implementation. I modified it to comply with the Torvalds three-levels-
of-indentation rule.

It will be used to keep track of buffer markers in the future. Those can
then be added by (custom) modes, or perhaps also by the user (like in
vim).