Add direction field to BufferMarker
Eventually, when buffer markers are kept track of by the buffer, they're
split into two binary trees: one after the "high selection marker" (the
greatest of sel_start and sel_finish), and one for markers before that
point. To ensure the least amount of marker invalidations, all markers
in the lower tree have direction 1, and all markers in the higher tree
have direction -1.