Add marker traversal logic
When the selections changes, markers can go from hi to lo or lo to hi.
This should now actually be implemented. Untested though (I know this is
the master branch, but basic editing still works and damn it: I'll
adhere to sensible branching and stability rules once the product is
actually stable enough that that shit matters).